do to ordertail

+++ BuildWeb/vegetable/src/main/java/com/product/entity/OrderDetail.java
@@ -14,7 +14,7 @@ public class OrderDetail {
 	   @Id
 	   @GeneratedValue(strategy = GenerationType.IDENTITY)
 	   private int id;
-	   private Double unitPrice;
+	   private Float unitPrice;
 	   private int quantity;
 	   private Double discount;
 	   private String size;
@@ -26,60 +26,60 @@ public class OrderDetail {
 	   @JoinColumn(name="productId")
 	   private Product product;
 
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
-
-	public Double getUnitPrice() {
-		return unitPrice;
-	}
-
-	public void setUnitPrice(Double unitPrice) {
-		this.unitPrice = unitPrice;
-	}
-
-	public int getQuantity() {
-		return quantity;
-	}
-
-	public void setQuantity(int quantity) {
-		this.quantity = quantity;
-	}
-
-	public Double getDiscount() {
-		return discount;
-	}
-
-	public void setDiscount(Double discount) {
-		this.discount = discount;
-	}
-
-	public String getSize() {
-		return size;
-	}
-
-	public void setSize(String size) {
-		this.size = size;
-	}
-
-	public Order getOrder() {
-		return order;
-	}
-
-	public void setOrder(Order order) {
-		this.order = order;
-	}
-
-	public Product getProduct() {
-		return product;
-	}
-
-	public void setProduct(Product product) {
-		this.product = product;
-	}
+		public int getId() {
+			return id;
+		}
+	
+		public void setId(int id) {
+			this.id = id;
+		}
+	
+		public Float getUnitPrice() {
+			return unitPrice;
+		}
+	
+		public void setUnitPrice(Float unitPrice) {
+			this.unitPrice = unitPrice;
+		}
+	
+		public int getQuantity() {
+			return quantity;
+		}
+	
+		public void setQuantity(int quantity) {
+			this.quantity = quantity;
+		}
+	
+		public Double getDiscount() {
+			return discount;
+		}
+	
+		public void setDiscount(Double discount) {
+			this.discount = discount;
+		}
+	
+		public String getSize() {
+			return size;
+		}
+	
+		public void setSize(String size) {
+			this.size = size;
+		}
+	
+		public Order getOrder() {
+			return order;
+		}
+	
+		public void setOrder(Order order) {
+			this.order = order;
+		}
+	
+		public Product getProduct() {
+			return product;
+		}
+	
+		public void setProduct(Product product) {
+			this.product = product;
+		}
 	   
 }
